+++ expenses/forms/__init__.py
@@ -4,6 +4,7 @@ from .payment import PaymentForm, ExpenseItemEditForm
 from .payee import PayeeForm
 from .budget import BudgetForm
 from .payment_method import PaymentMethodForm
+from .quick_expense import QuickExpenseForm
 
 # Make all form classes available when importing from expenses.forms
 __all__ = [
@@ -13,4 +14,5 @@ __all__ = [
     "PayeeForm",
     "BudgetForm",
     "PaymentMethodForm",
+    "QuickExpenseForm",
 ]